- Swirched to new meshing to work on for later - Cleared a bunch of warnings - More idk i forgot

+++ Minecraft Clone/src/Client/Client.cpp
@@ -66,7 +66,7 @@ void Client::Update() {
 
 	UpdateKeyPressSet();
 
-	MainLocalWorld.UpdateIO(GetKeyboardDump(), TestForKeyInputs(GLFW_KEY_LEFT_SHIFT), (float)cursormovementx, (float)cursormovementy, PressedLeft, PressedMiddle, PressedRight, Frametime);
+	MainLocalWorld.UpdateIO(GetKeyboardDump(), TestForKeyInputs(GLFW_KEY_LEFT_SHIFT), (float)cursormovementx, (float)cursormovementy, PressedLeft, PressedMiddle, PressedRight, (float)Frametime);
 	PressedLeft = false;
 	PressedMiddle = false;
 	PressedRight = false;
